Silence type_complexity lint in one files
The host state contains two tables with Arc & Mutex wrapping, which
clippy doesn't really like.

Although somewhat of an eyesore, prefactoring to fix this is silly
because we don't know what the server state will evolve to and the
actual types are fairly obvious (standard concurrency wrappers, tuples
of a relation on the inside).