gh workflow: use magic-nix-cache
Previously the plain cache action was used for coarse grained caching of
the entire /nix/ store, and cachix was used for fine grained caching,
requiring an API key for uploads.

The magic-nix-cache github action provides cachix like semantics, i.e.
are nix aware and per package, but using github's caching infrastructure
as the backing store, which is more local and has implicit access
control, removing a lot of boilerplate for what should be a performance
win.